// ClientApp/src/app/core/models/mappedmeter.model.ts
export interface IMappedMeter {
    MappedMeterId: number;
    BuildingId: number;
    BuildingName: string;
    PartnerId: number;
    PartnerName: string;
    BuildingServiceId: number;
    MeterNo: string;
    Description: string;
    UmfaDescription: string;
    ScadaSerial: string;
    ScadaDescription: string;
    RegisterType: string;
    TOUHeader: string;
    SupplyType: string;
    SupplyTo: string;
    Location: string;
    TOUId: number;
    RegisterTypeId: number;
}